Common TunnelResponseHeadersOrTrailers for UDP & TCP tunneling
