+++ envoy/http/BUILD
@@ -140,8 +140,10 @@ envoy_cc_library(
     deps = [
         ":header_formatter_interface",
         "//envoy/common:union_string",
+        "//envoy/stream_info:filter_state_interface",
         "//source/common/common:assert_lib",
         "//source/common/common:hash_lib",
+        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
     ],
 )
