+++ source/common/tcp_proxy/tcp_proxy.h
@@ -117,18 +117,9 @@ using TunnelingConfig =
     envoy::extensions::filters::network::tcp_proxy::v3::TcpProxy_TunnelingConfig;
 
 /**
- * Base class for both tunnel response headers and trailers.
- */
-class TunnelResponseHeadersOrTrailers : public StreamInfo::FilterState::Object {
-public:
-  ProtobufTypes::MessagePtr serializeAsProto() const override;
-  virtual const Http::HeaderMap& value() const PURE;
-};
-
-/**
  * Response headers for the tunneling connections.
  */
-class TunnelResponseHeaders : public TunnelResponseHeadersOrTrailers {
+class TunnelResponseHeaders : public Http::TunnelResponseHeadersOrTrailers {
 public:
   TunnelResponseHeaders(Http::ResponseHeaderMapPtr&& response_headers)
       : response_headers_(std::move(response_headers)) {}
@@ -142,7 +133,7 @@ private:
 /**
  * Response trailers for the tunneling connections.
  */
-class TunnelResponseTrailers : public TunnelResponseHeadersOrTrailers {
+class TunnelResponseTrailers : public Http::TunnelResponseHeadersOrTrailers {
 public:
   TunnelResponseTrailers(Http::ResponseTrailerMapPtr&& response_trailers)
       : response_trailers_(std::move(response_trailers)) {}
